create separate components for sections on landing page
